requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
pymediawiki==0.6.2
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
